refactor: move pickContext to its own module
